WinUI - Implement SettingsDialog
