+++ NickvisionApplication.WinUI/NickvisionApplication.WinUI.csproj
@@ -17,7 +17,7 @@
     <None Remove="Controls\StatusPage.xaml" />
     <None Remove="Controls\ViewStack.xaml" />
     <None Remove="Controls\ViewStackPage.xaml" />
-    <None Remove="Views\PreferencesPage.xaml" />
+    <None Remove="Views\SettingsDialog.xaml" />
   </ItemGroup>
 
   <ItemGroup>
@@ -54,12 +54,12 @@
 	  <Resource Include="Assets\org.nickvision.application.ico" />
   </ItemGroup>
   <ItemGroup>
-    <Page Update="Controls\ActionRow.xaml">
+    <Page Update="Views\SettingsDialog.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
   </ItemGroup>
   <ItemGroup>
-    <Page Update="Views\PreferencesPage.xaml">
+    <Page Update="Controls\ActionRow.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
   </ItemGroup>
