+++ NickvisionApplication.WinUI/Views/MainWindow.xaml
@@ -83,7 +83,43 @@
                 </MenuBarItem>
 
                 <MenuBarItem x:Name="MenuHelp">
+                    <MenuFlyoutItem x:Name="MenuChangelog" Click="Changelog">
+                        <MenuFlyoutItem.Icon>
+                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xEC24;"/>
+                        </MenuFlyoutItem.Icon>
+                    </MenuFlyoutItem>
 
+                    <MenuFlyoutSeparator/>
+
+                    <MenuFlyoutItem x:Name="MenuGitHubRepo" Click="GitHubRepo">
+                        <MenuFlyoutItem.Icon>
+                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xE1D3;"/>
+                        </MenuFlyoutItem.Icon>
+                    </MenuFlyoutItem>
+
+                    <MenuFlyoutItem x:Name="MenuReportABug" Click="ReportABug">
+                        <MenuFlyoutItem.Icon>
+                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xEBE8;"/>
+                        </MenuFlyoutItem.Icon>
+                    </MenuFlyoutItem>
+
+                    <MenuFlyoutItem x:Name="MenuDiscussions" Click="Discussions">
+                        <MenuFlyoutItem.Icon>
+                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xE15F;"/>
+                        </MenuFlyoutItem.Icon>
+                    </MenuFlyoutItem>
+
+                    <MenuFlyoutSeparator/>
+
+                    <MenuFlyoutItem x:Name="MenuAbout" Click="About">
+                        <MenuFlyoutItem.Icon>
+                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xE946;"/>
+                        </MenuFlyoutItem.Icon>
+
+                        <MenuFlyoutItem.KeyboardAccelerators>
+                            <KeyboardAccelerator Modifiers="None" Key="F1"/>
+                        </MenuFlyoutItem.KeyboardAccelerators>
+                    </MenuFlyoutItem>
                 </MenuBarItem>
             </MenuBar>
         </wct:DockPanel>
@@ -130,8 +166,6 @@
                             </Border>
                         </Grid>
                     </nickvision:ViewStackPage>
-
-                    <nickvision:ViewStackPage x:Name="PageSettings" PageName="Settings"/>
                 </nickvision:ViewStack.Pages>
             </nickvision:ViewStack>
