--- NickvisionApplication.WinUI/Views/PreferencesPage.xaml
@@ -1,98 +1,0 @@
-<UserControl
-    x:Class="NickvisionApplication.WinUI.Views.PreferencesPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:NickvisionApplication.WinUI.Views"
-    xmlns:nickvision="using:NickvisionApplication.WinUI.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:wct="using:CommunityToolkit.WinUI.UI.Controls"
-    xmlns:labs="using:CommunityToolkit.Labs.WinUI"
-    mc:Ignorable="d">
-
-    <Grid>
-        <TextBlock x:Name="LblTitle" Margin="30,30,0,0" Style="{ThemeResource TitleTextBlockStyle}"/>
-
-        <wct:DockPanel HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="0,90,0,10">
-            <StackPanel wct:DockPanel.Dock="Right" Margin="10,0,10,0">
-                <TextBlock x:Name="LblAbout" Style="{ThemeResource NavigationViewItemHeaderTextStyle}"/>
-
-                <TextBlock x:Name="LblDescription" Foreground="Gray"/>
-
-                <TextBlock x:Name="LblVersion"/>
-
-                <TextBlock Foreground="Gray" Text="(C) Nickvision 2021-2022&#xA;"/>
-
-                <HyperlinkButton Margin="-12,0,0,0" Click="Changelog">
-                    <HyperlinkButton.Content>
-                        <StackPanel Orientation="Horizontal" Spacing="6">
-                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xE8FD;"/>
-
-                            <TextBlock x:Name="LblBtnChangelog"/>
-                        </StackPanel>
-                    </HyperlinkButton.Content>
-                </HyperlinkButton>
-
-                <HyperlinkButton Margin="-12,0,0,0" Click="Credits">
-                    <HyperlinkButton.Content>
-                        <StackPanel Orientation="Horizontal" Spacing="6">
-                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xE125;"/>
-
-                            <TextBlock x:Name="LblBtnCredits"/>
-                        </StackPanel>
-                    </HyperlinkButton.Content>
-                </HyperlinkButton>
-
-                <HyperlinkButton Margin="-12,12,0,0" Click="GitHubRepo">
-                    <HyperlinkButton.Content>
-                        <StackPanel Orientation="Horizontal" Spacing="6">
-                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xE1D3;"/>
-
-                            <TextBlock x:Name="LblBtnGitHubRepo"/>
-                        </StackPanel>
-                    </HyperlinkButton.Content>
-                </HyperlinkButton>
-
-                <HyperlinkButton Margin="-12,0,0,0" Click="ReportABug">
-                    <HyperlinkButton.Content>
-                        <StackPanel Orientation="Horizontal" Spacing="6">
-                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xEBE8;"/>
-
-                            <TextBlock x:Name="LblBtnReportABug"/>
-                        </StackPanel>
-                    </HyperlinkButton.Content>
-                </HyperlinkButton>
-
-                <HyperlinkButton Margin="-12,0,0,0" Click="Discussions">
-                    <HyperlinkButton.Content>
-                        <StackPanel Orientation="Horizontal" Spacing="6">
-                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xE15F;"/>
-
-                            <TextBlock x:Name="LblBtnDiscussions"/>
-                        </StackPanel>
-                    </HyperlinkButton.Content>
-                </HyperlinkButton>
-            </StackPanel>
-
-            <ScrollViewer wct:DockPanel.Dock="Left">
-                <StackPanel Margin="20,0,20,0" Spacing="6">
-                    <labs:SettingsExpander x:Name="CardUserInterface">
-                        <labs:SettingsExpander.HeaderIcon>
-                            <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xE771;"/>
-                        </labs:SettingsExpander.HeaderIcon>
-
-                        <labs:SettingsExpander.Items>
-                            <labs:SettingsCard x:Name="CardTheme">
-                                <labs:SettingsCard.HeaderIcon>
-                                    <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" Glyph="&#xE790;"/>
-                                </labs:SettingsCard.HeaderIcon>
-
-                                <ComboBox x:Name="CmbTheme" SelectionChanged="CmbTheme_SelectionChanged"/>
-                            </labs:SettingsCard>
-                        </labs:SettingsExpander.Items>
-                    </labs:SettingsExpander>
-                </StackPanel>
-            </ScrollViewer>
-        </wct:DockPanel>
-    </Grid>
-</UserControl>
